Fix pre-commit check error and address review comments

+++ config/chassis_modules.py
@@ -3,8 +3,11 @@
 import click
 import time
 import re
-from swsscommon.swsscommon import SonicV2Connector
+import subprocess
 import utilities_common.cli as clicommon
+
+TIMEOUT_SECS = 10
+
 
 #
 # 'chassis_modules' group ('config chassis_modules ...')
@@ -19,6 +22,7 @@ def modules():
     """Configure chassis modules"""
     pass
 
+
 def get_config_module_state(db, chassis_module_name):
     config_db = db.cfgdb
     fvs = config_db.get_entry('CHASSIS_MODULE', chassis_module_name)
@@ -27,21 +31,21 @@ def get_config_module_state(db, chassis_module_name):
     else:
         return fvs['admin_status']
 
-TIMEOUT_SECS = 10
 
 #
-# Name: get_config_module_state_timeout
+# Name: check_config_module_state_with_timeout
 # return: True: timeout, False: not timeout
 #
-def get_config_module_state_timeout(ctx, db, chassis_module_name, state):
+def check_config_module_state_with_timeout(ctx, db, chassis_module_name, state):
     counter = 0
-    while  get_config_module_state(db, chassis_module_name) != state:
+    while get_config_module_state(db, chassis_module_name) != state:
         time.sleep(1)
         counter += 1
         if counter >= TIMEOUT_SECS:
             ctx.fail("get_config_module_state {} timeout".format(chassis_module_name))
             return True
     return False
+
 
 def get_asic_list_from_db(chassisdb, chassis_module_name):
     asic_list = []
@@ -53,11 +57,12 @@ def get_asic_list_from_db(chassisdb, chassis_module_name):
             asic_list.append(asic_id)
     return asic_list
 
+
 #
 # Syntax: fabric_module_set_admin_status <chassis_module_name> <'up'/'down'>
 #
-def fabric_module_set_admin_status(chassis_module_name, state):
-    chassisdb = SonicV2Connector(host="127.0.0.1")
+def fabric_module_set_admin_status(db, chassis_module_name, state):
+    chassisdb = db.db
     chassisdb.connect("CHASSIS_STATE_DB")
     asic_list = get_asic_list_from_db(chassisdb, chassis_module_name)
 
@@ -66,16 +71,30 @@ def fabric_module_set_admin_status(chassis_module_name, state):
 
     if state == "down":
         for asic in asic_list:
-            click.echo("Stop swss@{} and syncd@{} ...".format(asic, asic))
+            click.echo("Stop swss@{} and syncd@{} services".format(asic, asic))
             clicommon.run_command('sudo systemctl stop swss@{}.service'.format(asic))
-            # wait for service is down
-            time.sleep(2)
-            chassisdb.delete("CHASSIS_STATE_DB","CHASSIS_FABRIC_ASIC_TABLE|asic" + str(asic))
-            click.echo("Start swss@{} and syncd@{} ...".format(asic, asic))
-            clicommon.run_command('sudo systemctl start swss@{}.service'.format(asic))
-    else:
+
+        is_active = subprocess.call(["systemctl", "is-active", "--quiet", "swss@{}.service".format(asic)])
+
+        if is_active == 0:  # zero active,  non-zero, inactive
+            click.echo("stop swss@{} and syncd@{} services failed".format(asic, asic))
+            return
+
+        click.echo("Delete related CAHSSIS_FABRIC_ASIC_TABLE entries")
+
         for asic in asic_list:
-            click.echo("Start swss@{} and syncd@{} ...".format(asic, asic))
+            chassisdb.delete("CHASSIS_STATE_DB", "CHASSIS_FABRIC_ASIC_TABLE|asic" + str(asic))
+
+        # Start the services in case of the users just execute issue command "systemctl stop swss@/syncd@"
+        # without bring down the hardware
+        for asic in asic_list:
+            click.echo("Start swss@{} and syncd@{} services".format(asic, asic))
+            # To address systemd service restart limit by resetting the count
+            clicommon.run_command('sudo systemctl reset-failed swss@{}.service'.format(asic))
+            clicommon.run_command('sudo systemctl start swss@{}.service'.format(asic))
+    elif state == "up":
+        for asic in asic_list:
+            click.echo("Start swss@{} and syncd@{} services".format(asic, asic))
             clicommon.run_command('sudo systemctl start swss@{}.service'.format(asic))
 
 #
@@ -94,7 +113,7 @@ def shutdown_chassis_module(db, chassis_module_name):
        not chassis_module_name.startswith("FABRIC-CARD"):
         ctx.fail("'module_name' has to begin with 'SUPERVISOR', 'LINE-CARD' or 'FABRIC-CARD'")
 
-    #To avoid duplicate operation
+    # To avoid duplicate operation
     if get_config_module_state(db, chassis_module_name) == 'down':
         click.echo("Module {} is already in down state".format(chassis_module_name))
         return
@@ -103,8 +122,8 @@ def shutdown_chassis_module(db, chassis_module_name):
     fvs = {'admin_status': 'down'}
     config_db.set_entry('CHASSIS_MODULE', chassis_module_name, fvs)
     if chassis_module_name.startswith("FABRIC-CARD"):
-        if not get_config_module_state_timeout(ctx, db, chassis_module_name, 'down'):
-            fabric_module_set_admin_status(chassis_module_name, 'down')
+        if not check_config_module_state_with_timeout(ctx, db, chassis_module_name, 'down'):
+            fabric_module_set_admin_status(db, chassis_module_name, 'down')
 
 #
 # 'startup' subcommand ('config chassis_modules startup ...')
@@ -117,7 +136,7 @@ def startup_chassis_module(db, chassis_module_name):
     config_db = db.cfgdb
     ctx = click.get_current_context()
 
-    #To avoid duplicate operation
+    # To avoid duplicate operation
     if get_config_module_state(db, chassis_module_name) == 'up':
         click.echo("Module {} is already set to up state".format(chassis_module_name))
         return
@@ -125,5 +144,5 @@ def startup_chassis_module(db, chassis_module_name):
     click.echo("Starting up chassis module {}".format(chassis_module_name))
     config_db.set_entry('CHASSIS_MODULE', chassis_module_name, None)
     if chassis_module_name.startswith("FABRIC-CARD"):
-        if not get_config_module_state_timeout(ctx, db, chassis_module_name, 'up'):
-            fabric_module_set_admin_status(chassis_module_name, 'up')
+        if not check_config_module_state_with_timeout(ctx, db, chassis_module_name, 'up'):
+            fabric_module_set_admin_status(db, chassis_module_name, 'up')
